requirements.txt
matplotlib
Pillow
pandas
numpy
reportlab
openpyxl